Updating code for CRUD and Swagger implementation

+++ src/main/java/com/retail/retailChain/repository/RetailRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.retail.retailChain.entity.RetailStore;
 
 
-public interface RetailRepository extends JpaRepository<RetailStore,Integer>,RetailRepositoryCustom {
+public interface RetailRepository extends JpaRepository<RetailStore,Long>,RetailRepositoryCustom {
 	//public RetailStore findByProductName();
 
 }
